+++ src/main/java/com/example/easybooking/shop/repository/ShopRepository.java
@@ -1,5 +1,6 @@
-package com.example.easybooking.shop;
+package com.example.easybooking.shop.repository;
 
+import com.example.easybooking.shop.domain.Shop;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ShopRepository extends JpaRepository<Shop, Long> {
