+++ src/main/java/com/example/easybooking/slot/Slot.java
@@ -1,6 +1,6 @@
 package com.example.easybooking.slot;
 
-import com.example.easybooking.shop.Shop;
+import com.example.easybooking.shop.domain.Shop;
 import jakarta.persistence.*;
 import lombok.Getter;
